chore: split upgrade.go into 3 files (#939)
* chore: split upgrade.go into 3 files


* update: code review


* revert: keep return error for the case of readfile etc


* Update pkg/cluster/cluster_config.go


---------
