+++ pkg/cluster/cluster_config.go
@@ -4,7 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"os"
 
+	ofapi "github.com/operator-framework/api/pkg/operators/v1alpha1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -28,4 +30,27 @@ func GetDomain(c client.Client) (string, error) {
 	}
 
 	return domain, err
+}
+
+func GetOperatorNamespace() (string, error) {
+	data, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
+	return string(data), err
+}
+
+// GetClusterServiceVersion retries the clusterserviceversions available in the operator namespace.
+func GetClusterServiceVersion(ctx context.Context, c client.Client, watchNameSpace string) (*ofapi.ClusterServiceVersion, error) {
+	clusterServiceVersionList := &ofapi.ClusterServiceVersionList{}
+	if err := c.List(ctx, clusterServiceVersionList, client.InNamespace(watchNameSpace)); err != nil {
+		return nil, fmt.Errorf("failed listign cluster service versions: %w", err)
+	}
+
+	for _, csv := range clusterServiceVersionList.Items {
+		for _, operatorCR := range csv.Spec.CustomResourceDefinitions.Owned {
+			if operatorCR.Kind == "DataScienceCluster" {
+				return &csv, nil
+			}
+		}
+	}
+
+	return nil, nil
 }
